refactor: organize platform handling logic
